requirements.txt
alembic==0.8.10
awesome-slugify==1.6.5
bcrypt==3.1.3
blinker==1.3
cffi==1.9.1
click==6.7
dominate==2.3.1
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Misaka==0.4.1
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy-Redux==0.7.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
misaka==2.1.0
pycparser==2.17
python-editor==1.0.3
regex==2017.2.8
six==1.10.0
SQLAlchemy==1.1.5
Unidecode==0.4.20
visitor==0.1.3
Werkzeug==0.11.15
Whoosh==2.6.0
WTForms==2.1
alembic==0.8.10
awesome-slugify==1.6.5
bcrypt==3.1.3
blinker==1.3
cffi==1.9.1
click==6.7
dominate==2.3.1
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Misaka==0.4.1
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy-Redux==0.7.1
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
misaka==2.1.0
pycparser==2.17
python-editor==1.0.3
regex==2017.2.8
six==1.10.0
SQLAlchemy==1.1.5
Unidecode==0.4.20
visitor==0.1.3
Werkzeug==0.11.15
Whoosh==2.6.0
WTForms==2.1
